requirements.txt
Flask
notion
slackeventsapi
slackclient
python-dotenv
pprint
peewee